- Changed WeaponState from 8-bit to 16-bit integer. - Because the flags for WF_USER#OK are 256 on up, this is required in order to work.

+++ src/d_player.h
@@ -409,7 +409,7 @@ public:
 	int			lastkilltime;			// [RH] For multikills
 	BYTE		multicount;
 	BYTE		spreecount;				// [RH] Keep track of killing sprees
-	BYTE		WeaponState;
+	WORD		WeaponState;
 
 	AWeapon	   *ReadyWeapon;
 	AWeapon	   *PendingWeapon;			// WP_NOCHANGE if not changing
